update restful api
