+++ TODO.md
@@ -3,6 +3,11 @@ Feature
 [] create a frontend, transfer to a restful service
 [] Add Readme with steps to start app
 [] create Restful documentation
+  - test basic function (PostController without auth) ok
+  - use interface to map out method for a service or controller
+  - check how to add auth controller
+  - check best practice to add auth check to all auth controllers
+  - 2 steps to build rest api -> build by function -> build higher level api by UI requirements
 Bugfix
 [] map a fixed port to postgres for docker
 [] loadRepository doesn't work
