+++ minithread-be/src/main/java/com/en/training/minithread/controllers/wsdtos/AccountWsDTO.java
@@ -1,0 +1,25 @@
+package com.en.training.minithread.controllers.wsdtos;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class AccountWsDTO {
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
